Fix security error on duplication and fix vuln in auth
