+++ frontend/src/app/api/auth/register/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
-import { registerUser, loginWithPassword } from "@/lib/keycloakService";
 
-const BACKEND_URL = process.env.API_URL ?? "http://localhost:3001";
+import { createAuthResponse } from "@/lib/authSession";
+import { loginWithPassword, registerUser } from "@/lib/keycloakService";
 
 export async function POST(req: NextRequest) {
   const body = await req.json() as Record<string, string>;
@@ -17,38 +17,7 @@ export async function POST(req: NextRequest) {
   try {
     await registerUser(username, email, password, firstName, lastName);
     const tokens = await loginWithPassword(username, password);
-
-    const res = NextResponse.json({ success: true });
-
-    res.cookies.set("access_token", tokens.access_token, {
-      httpOnly: true,
-      secure: process.env.NODE_ENV === "production",
-      sameSite: "lax",
-      maxAge: tokens.expires_in,
-      path: "/",
-    });
-
-    res.cookies.set("refresh_token", tokens.refresh_token, {
-      httpOnly: true,
-      secure: process.env.NODE_ENV === "production",
-      sameSite: "lax",
-      maxAge: tokens.refresh_expires_in,
-      path: "/",
-    });
-
-    const provisionRes = await fetch(`${BACKEND_URL}/api/v1/users/me`, {
-      headers: { Authorization: `Bearer ${tokens.access_token}` },
-    }).catch((err: unknown) => {
-      console.error("[register] backend unreachable during provisioning:", err);
-      return null;
-    });
-
-    if (provisionRes && !provisionRes.ok) {
-      const body = await provisionRes.text().catch(() => "");
-      console.error(`[register] provisioning returned ${provisionRes.status}:`, body);
-    }
-
-    return res;
+    return createAuthResponse(tokens);
   } catch (err) {
     const message = err instanceof Error ? err.message : "Registration failed";
     const status = message.includes("already exists") ? 409 : 400;
